requirements.txt
multitasking~=0.0.9
pillow>=8.2.0
requests~=2.25.1
spotipy~=2.19.0